Start work on custom OAI endpoints
